// generated/src/app/api-models/member-contact-arc.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class MemberContactArc {
  seqMembId: number;
  contactSource: string;
  changeDateTime: Date;
  primaryDistributionMethod: string;
  emailId: string;
  faxNumber: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}